Split weather tests into two files, one for programmatic tests and the other for scenario tests
